+++ openpipe/tests/test_libraries.py
@@ -14,6 +14,9 @@ def test_remote_lib():
         exit_code, exit_message = pipeline_run("openpipe/tests/libraries_auto_net.yaml")
     except ModuleNotFoundError:
         pass
+    except SystemExit as se:
+        if se.code == 0:
+            raise Exception("Loaded file without net enabled")
     else:
         raise Exception("Loaded file without net enabled")
     environ["OPENPIPE_AUTO_NETWORK"] = "True"