+++ CHANGELOG.md
@@ -15,3 +15,4 @@ This minor release brings the following improvements:
 * Add 'read from random' action to fetch random numbers (closes #19)
 * Add 'queue' action to queue input items into a list (closes #20)
 * Add 'join' action to join elements from an input list item (closes #21)
+* Add action 'tag as' to use the input item as a tag key value (closes #22)